Production

+++ services/projectService.js
@@ -1,13 +1,10 @@
 import API from "./api";
 import * as youtubeService from '../services/youtubeService';
+import * as tagService from './tagService';
 
 export const findAll = async (page, size, token) => {
   API.defaults.headers.common["Authorization"] = token;
-  const projects = await API.get(`/projects?page=${page}&size=${size}`);
-  projects.forEach((project) => {
-    project.previewImage = `${process.env.NEXT_PUBLIC_HOST_BACKEND}/images/projects/${project.id}/${project.previewImage}`;
-  });
-  return projects;
+  return await API.get(`/projects?page=${page}&size=${size}`);
 };
 
 
@@ -16,40 +13,68 @@ export const getById = async (id, token) => {
   const project = await API.get(`/projects/${id}`);
   let purchased = await isPurchased(project, token);
   project.purchased = purchased != undefined; 
-  project.previewImage = `${process.env.NEXT_PUBLIC_HOST_BACKEND}/images/projects/${project.id}/${project.previewImage}`;
   return project;
 };
 
-export const addProject = async (project, token, id) => {
-  delete project.previewImage;
-  delete project.images;
-  delete project.files;
+export const addProject = async (project, token) => {
+  API.defaults.headers.common["Authorization"] = token;
+  const copyProject = Object.assign({}, project);
+  delete copyProject.previewImage;
+  delete copyProject.images;
+  delete copyProject.file;
   try {
-    project.videoPath = youtubeService.getIdVideo(project.videoPath);
-    API.defaults.headers.common["Authorization"] = token;
-    return await API.post(`/projects?professional=${id}`, project);
+    copyProject.videoPath = youtubeService.getIdVideo(project.videoPath);
   } catch (err) {
     throw new Error(`Error, link to video invalid: ${err}`);
   }
+
+  const projectUploaded = await API.post(`/projects`, copyProject);
+  
+  if (project.previewImage) {
+    await addPreviewImage(project.previewImage, projectUploaded.id, token);
+    projectUploaded.previewImage = URL.createObjectURL(project.previewImage);
+  }
+  
+  if (project.images.length > 0) {
+    await addImages(project.images, projectUploaded.id, token);
+  }
+    
+  if (project.file)
+    await addFile(project.file, projectUploaded.id, token);
+    
+  return projectUploaded;
 };
 
 export const addPreviewImage = async (image, projectId, token) => {
   API.defaults.headers.common["Authorization"] = token;
   const imageData = new FormData();
   imageData.append("imageFile", image);
-  imageData.append("project", projectId);
-  return await API.post(`/images/project/preview`, imageData);
+  return await API.post(`/images/projects/${projectId}/preview`, imageData);
 };
 
 export const addImages = async (images, projectId, token) => {
-  API.defaults.headers.common["Authorization"] = token;
-  
-  images.forEach(async (image) => {
-    const imageData = new FormData();
-    imageData.append("imageFile", image);
-    imageData.append("project", projectId);
-    imageData.append("tags", []);
-    await API.post(`/images`, imageData);
+  await addImagesRecursive(Array.from(images), projectId, token);
+};
+
+export const addImagesRecursive = async (images, projectId, token) => {
+  const image = images.shift();
+  await addImage(image, projectId, token);
+  if (images.length > 0) {
+    await addImagesRecursive(images, projectId, token);
+  }
+};
+
+export const addImage = async (image, projectId, token) => {
+  const imageData = new FormData();
+  const tags = tagService.getTags(image.tags);
+  imageData.append("imageFile", image);
+  imageData.append("tags", tags);
+  image.uploading = true;
+  return await API.post(`/images/projects/${projectId}`, imageData, {
+    onUploadProgress: progressEvent => {
+      const progress = Math.round(progressEvent.loaded / progressEvent.total * 100);
+      image.setProgress(progress);
+    }
   });
 };
 
@@ -64,7 +89,6 @@ export const edit = async (project, token) => {
   delete project.id;
   delete project.images;
   let projectEdited = await API.put(`/projects/${id}`, project);
-  projectEdited.previewImage = `${process.env.NEXT_PUBLIC_HOST_BACKEND}/images/projects/${id}/${projectEdited.previewImage}`;
   if (previewImage) {
     await addPreviewImage(previewImage, id, token);
     projectEdited.previewImage = URL.createObjectURL(previewImage);
@@ -100,21 +124,29 @@ const removeAndAddImages = async (images, id, token) => {
   
   for (let i = 0; i < images.length; i++) {
     let img = images[i];
-    if (img.added && img.remove) {
+    if (img.added && img.remove) { //is so that the image is already in the database and wants to delete it
       await API.delete(`/images/projects/${id}/${img.name}`);
-    } else if (!img.added) {
+    } else if (!img.added) { //a new image is not in the database 
       const imageData = new FormData();
       imageData.append("imageFile", img);
-      imageData.append("project", id);
       imageData.append("tags", []);
-      await API.post(`/images`, imageData);
+      await API.post(`/images/projects/${id}`, imageData);
       const path = URL.createObjectURL(img);
       newImages.push({ path });
-    } else {
+    } else if (img.added) { //this in the database does not want to delete it but wants to update the tags or can not be the same
+      await editTags(img, token);
       newImages.push({ path: img.path });
     }
   };
   return newImages;
+};
+
+export const editTags = async (image, token) => {
+  API.defaults.headers.common["Authorization"] = token;
+  const imageData = new FormData();
+  const tags = tagService.getTags(image.tags);
+  imageData.append("tags", tags);
+  return await API.put(`/images/${image.id}/projects/tags`, imageData);
 };
 
 export const getPurchasedProjects = async (token) => {
@@ -128,4 +160,12 @@ export const isPurchased = async (project, token) => {
   return purchasedProjects.find(purchased => purchased.id == project.id);
 };
 
-export default {findAll, getById, addProject, addPreviewImage, addImages, edit, addFile, download, removeAndAddImages,getPurchasedProjects, isPurchased}
+export const findAllByProfessionalId = async (professionalId, page, size, token) => {
+  API.defaults.headers.common["Authorization"] = token;
+  return await API.get(`/projects/professionals/${professionalId}?page=${page}&size=${size}`);
+  
+};
+
+export const findByNameAndActive = async (name, active, page, size) => {
+  return await API.get(`/projects/name/${name}/active/${active}?page=${page}&size=${size}`);
+}
